+++ source/Calamari/Calamari.csproj
@@ -86,7 +86,7 @@
     </None>
   </ItemGroup>
   <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0' ">
-x    <PackageReference Include="KubernetesClient" Version="15.0.1" />
+    <PackageReference Include="KubernetesClient" Version="15.0.1" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="ArgoCD\Http\" />
